pyproject.toml
[project]
name = "agentforce-mcp"
version = "0.1.0"
description = "MCP server for Salesforce Agentforce Agent API"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp[cli]>=1.26.0",
    "httpx>=0.28.1",
    "starlette>=0.45.0",
    "uvicorn>=0.34.0",
    "python-dotenv>=1.0.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=0.25.0",
]

[tool.pytest.ini_options]
asyncio_mode = "strict"
